plot, loglog: added parentheses around let in 'if marker'
In plot, because the let in scope of if wasn't wrapped in
parentheses, the let extended to the rest of the closure,
and nothing else got executed if the if test failed. It's
also clear from the indentation that this isn't intended.
Without this change, plot settings didn't get restored
to their previous values at the end of the closure, which
caused some of them to carry over to the next closure
when it was executed.

It seems clear that the same problem exists in loglog, so I
fised that, too, although I have not tested loglog.  I get
an error when I try to run it. Nevertheless it seems obvious
that the change is needed.

(I haven't noticed other functions in owl_plot.ml that
have this problem.  I searched through all 'if' expressions
that use 'marker'.  Difficult to search for an if followed
by a let on the next line.)