Stats_dist: don't use balance_last for multinomial distribution (#347)
As described in #347, balance_last raises exceptions in case where
numerical inaccuracy prevent summation to 1 of input
probabilities. The comments there argue that not only the
transformation performed by balance_last, but is also not really
necessary. See notably this comment:

As I see it, owl's current implementation is very close to the one in
the GSL. It computes the sum of input probabilities (norm variable),
and it seems to me that it doesn't matter that these sum to 1. Playing
with gsl confirms it:

val rng : Gsl.Rng.t = <abstr>
- : int array = [|33251; 66749|]

This seems very natural to me, and much less confusing and error prone
than ignoring the last element of the probability vector.

So in summary, it would be enough to remove the call to balance_last
in multinomial_rvs:

let multinomial_rvs n ~p =
  let k = Array.length p in
  let _p = Genarray.create float64 c_layout [|k|] in
  let _s = Genarray.create int32 c_layout [|k|] in
  Owl_utils.Array.balance_last 1. p |>
  Array.iteri (fun i a -> Genarray.set _p [|i|] a);
  _multinomial_rvs ~k ~n ~p:_p _s;
  Array.init k (fun i -> Genarray.get _s [|i|] |> Int32.to_int)

The same is true IMO for the pdf:

- : float = 0.273129096174363206
- : float = 0.273129096174363706