Do not create .owl folder when loading owl library
I believe that libraries should not come with unexpected surprises, in particular I think that the main owl library should not create files in the system and instead file creation should be triggered by the user action (like downloading something from the zoo or a dataset). The creation of .owl/dataset and .owl/zoo has been moved to the zoo cli and the Dataset.load_data_path functions instead.

This change will also allow to run tests (at least some of them) inside
the opam sandboxed envirionment. In fact this is how I found out about
this.

Note that checking the existence of a folder and then creating it is racy (imagine you are downloading something in the zoo in a terminal and getting two datasets from two different instances of utop or jupyter).

Signed-off-by: Marcello Seri <marcello.seri@gmail.com>